vridge_back/pyproject.toml
[tool.poetry]
name = "vridge-back"
version = "0.1.0"
description = ""
authors = ["Jisu Park <wltn8223@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "vridge_back"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.1.0"
requests = "^2.31.0"
psycopg2-binary = "^2.9.6"
sentry-sdk = "^1.26.0"
moviepy = "^1.0.3"
channels = "^4.0.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
black = "^23.7.0"
whitenoise = "^6.5.0"
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
